% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_markers.R
\name{calc_maf}
\alias{calc_maf}
\title{Calculate minor allele frequency}
\usage{
calc_maf(x, check.matrix = TRUE)
}
\arguments{
\item{x}{An n x p marker matrix of n individuals and p markers coded as -1, 0, and 1
for homozygous alternate, heterozygous, and homozygous reference.}

\item{check.matrix}{Logical. Should the marker matrix 'x' be checked? Use check.matrix = FALSE
if 'x' contains imputed decimal genotypes.}
}
\description{
Calculate minor allele frequency
}
