% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_markers.R
\name{prune_LD}
\alias{prune_LD}
\title{Prune SNPs based on linkage disequilibrium}
\usage{
prune_LD(x, r2.max = 0.8, check.matrix = TRUE)
}
\arguments{
\item{x}{An n x p marker matrix of n individuals and p markers coded as -1, 0, and 1
for homozygous alternate, heterozygous, and homozygous reference.}

\item{check.matrix}{Logical. Should the marker matrix 'x' be checked? Use check.matrix = FALSE
if 'x' contains imputed decimal genotypes.}

\item{r2max}{The maximum acceptable value of r2 (i.e. LD) between any two markers.}
}
\value{
A marker matrix without markers in high LD
}
\description{
Prune SNPs based on linkage disequilibrium
}
